implement new records syncing; still needs housecleaning
