+++ app/assets/javascripts/live_record/model/relation.coffee
@@ -1,0 +1,8 @@
+LiveRecord.Model.Relation = (Model) ->
+	this
+
+LiveRecord.Model.Relation.prototype.where = ->
+	this
+
+LiveRecord.Model.Relation.prototype.to_a = ->
+
