+++ lib/live_record.rb
@@ -6,6 +6,7 @@ Dir[__dir__ + '/live_record/*.rb'].each {|file| require file }
 Dir[__dir__ + '/live_record/model/*.rb'].each {|file| require file }
 Dir[__dir__ + '/live_record/channel/*.rb'].each {|file| require file }
 Dir[__dir__ + '/live_record/generators/*.rb'].each {|file| require file }
+Dir[__dir__ + '/live_record/overrides/*.rb'].each {|file| require file }
 
 module LiveRecord
 end
