+++ lib/live_record/model/callbacks.rb
@@ -7,6 +7,7 @@ module LiveRecord
         before_update :__live_record_reference_changed_attributes__
         after_update_commit :__live_record_broadcast_record_update__
         after_destroy_commit :__live_record_broadcast_record_destroy__
+        after_create_commit :__live_record_broadcast_record_create__
 
         def self.live_record_whitelisted_attributes(record, current_user)
           []
@@ -22,13 +23,18 @@ module LiveRecord
           included_attributes = attributes.slice(*@_live_record_changed_attributes)
           @_live_record_changed_attributes = nil
           message_data = { 'action' => 'update', 'attributes' => included_attributes }
-          LiveRecordChannel.broadcast_to(self, message_data)
+          LiveRecord::ChangesChannel.broadcast_to(self, message_data)
           LiveRecordUpdate.create!(recordable_type: self.class, recordable_id: self.id, created_at: DateTime.now)
         end
 
         def __live_record_broadcast_record_destroy__
           message_data = { 'action' => 'destroy' }
-          LiveRecordChannel.broadcast_to(self, message_data)
+          LiveRecord::ChangesChannel.broadcast_to(self, message_data)
+        end
+
+        def __live_record_broadcast_record_create__
+          message_data = { 'action' => 'create', 'attributes' => attributes }
+          ActionCable.server.broadcast "live_record_publications_#{self.class.name.underscore}", message_data
         end
       end
     end
