+++ spec/internal/app/channels/application_cable/connection.rb
@@ -1,7 +1,7 @@
 module ApplicationCable
   class Connection < ActionCable::Connection::Base
   	identified_by :current_user
-
+  	
   	def current_user
   	end
   end
