--- spec/internal/app/channels/live_record_channel.rb
@@ -1,3 +1,0 @@
-class LiveRecordChannel < ApplicationCable::Channel
-  include LiveRecord::Channel::Implement
-end
