+++ spec/internal/db/schema.rb
@@ -10,7 +10,9 @@ ActiveRecord::Schema.define do
   create_table "posts", force: :cascade do |t|
     t.string "title"
     t.text "content"
+    t.boolean "is_enabled"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
+    t.index ["is_enabled"], name: "index_posts_on_is_enabled"
   end
 end
